+++ dotnet/Models/RetrieveTransaction.cs
@@ -76,7 +76,13 @@ namespace Cybersource.Models
     public class ApplicationInformation
     {
         [JsonProperty("reasonCode")]
-        public long ReasonCode { get; set; }
+        public string ReasonCode { get; set; }
+
+        [JsonProperty("rCode")]
+        public string RCode { get; set; }
+
+        [JsonProperty("rFlag")]
+        public string RFlag { get; set; }
 
         [JsonProperty("applications")]
         public List<Application> Applications { get; set; }
@@ -100,7 +106,10 @@ namespace Cybersource.Models
         public string RMessage { get; set; }
 
         [JsonProperty("returnCode")]
-        public long ReturnCode { get; set; }
+        public string ReturnCode { get; set; }
+
+        [JsonProperty("reconciliationId")]
+        public string ReconciliationId { get; set; }
     }
 
     [JsonObject(ItemNullValueHandling = NullValueHandling.Ignore)]
@@ -282,12 +291,36 @@ namespace Cybersource.Models
 
         [JsonProperty("proxyPan")]
         public string ProxyPan { get; set; }
+
+        [JsonProperty("signedParesStatusReason")]
+        public string SignedParesStatusReason { get; set; }
+
+        [JsonProperty("acsReferenceNumber")]
+        public string AcsReferenceNumber { get; set; }
+
+        [JsonProperty("ucafCollectionIndicator")]
+        public string UcafCollectionIndicator { get; set; }
+
+        [JsonProperty("directoryServerTransactionId")]
+        public string DirectoryServerTransactionId { get; set; }
+
+        [JsonProperty("threeDSServerTransactionId")]
+        public Guid ThreeDsServerTransactionId { get; set; }
+
+        [JsonProperty("acsOperatorID")]
+        public string AcsOperatorId { get; set; }
+
+        [JsonProperty("acsTransactionId")]
+        public Guid AcsTransactionId { get; set; }
     }
 
     public partial class StrongAuthentication
     {
         [JsonProperty("authenticationIndicator")]
         public string AuthenticationIndicator { get; set; }
+
+        [JsonProperty("OutageExemptionIndicator")]
+        public string OutageExemptionIndicator { get; set; }
     }
 
     public class Profile
@@ -308,7 +341,8 @@ namespace Cybersource.Models
         {
             StrongAuthentication = new StrongAuthentication
             {
-                AuthenticationIndicator = consumerAuthenticationInformation.StrongAuthentication != null ? consumerAuthenticationInformation.StrongAuthentication.AuthenticationIndicator : null
+                AuthenticationIndicator = consumerAuthenticationInformation.StrongAuthentication != null ? consumerAuthenticationInformation.StrongAuthentication.AuthenticationIndicator : null,
+                OutageExemptionIndicator = consumerAuthenticationInformation.StrongAuthentication != null ? consumerAuthenticationInformation.StrongAuthentication.OutageExemptionIndicator : null
             };
 
             AccessToken = consumerAuthenticationInformation.AccessToken;